backend/pyproject.toml
[tool.poetry]
authors = ["Oleksandr Korol <zibertua@gmail.com>"]
description = ""
license = "MIT"
name = "backend"
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
Django = "^4.2.4"
django-cors-headers = "^4.2.0"
django-filter = "^23.2"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = {extras = ["crypto"], version = "^5.3.0"}
drf-spectacular = "^0.26.4"
python = "3.11"
django-extensions = "^3.2.3"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
ruff = "^0.0.285"
mixer = "^7.2.2"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.ruff]
ignore = ["RUF005"]
line-length = 120
select = ["E", "F", "W", "I", "C4", "Q", "RUF", "PIE", "COM"]
# Exclude a variety of commonly ignored directories.
exclude = [
  ".git",
  ".hg",
  ".mypy_cache",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  "__pypackages__",
  "_build",
  "buck-out",
  "migrations",
  "dist",
  "venv",
  ".venv",
  ".conf",
  ".data",
  ".env",
  "asgi.py",
]
per-file-ignores = {}
# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
target-version = "py311"
[tool.ruff.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10
[tool.black]
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | ^.*\b(migrations)\b.*$
)/
'''
line-length = 120
target-version = ["py311"]
